+++ src/runtime/symbol_table.py
@@ -70,7 +70,7 @@ class SymbolTable:
         """Return the name in the symbol table that is the closest to 'name'. Return None if there is no close match."""
         if len(self.symbols) == 0 or name == "":
             return None
-        min_best_match = 0.3
+        min_best_match = 0.5
         best_match = min_best_match
         best_match_name = ""
         list_to_check = list(self.symbols.keys())